Swing_Income_Generator/requirements.txt

---

## 📄 `requirements.txt`

```txt
pandas
pyyaml
matplotlib